Refactor: Split React UI into separate modules
This commit modularizes the React application by breaking down the main App.js into smaller, more manageable components.

Key changes:
- Created a `src/components` directory for new UI components.
- Moved data fetching logic (`fetchData`) to `src/api.js`.
- Consolidated utility functions into `src/utils.js`.
- Extracted Stock, News, Sports, and Weather sections into their own components:
    - `src/components/StockInfo.js`
    - `src/components/NewsFeed.js`
    - `src/components/Sports.js`
    - `src/components/WeatherInfo.js`
- Extracted all modal dialogs into separate components:
    - `src/components/ManageStocksModal.js`
    - `src/components/ManageSportsModal.js`
    - `src/components/ManageRssModal.js`
    - `src/components/ManageWeatherModal.js`
- Cleaned up `App.js` to primarily handle state, orchestrate data fetching, and render the new components.

This refactoring aims to improve code clarity, maintainability, and separation of concerns.
