+++ src/App.js
@@ -2,22 +2,15 @@ import React, { useState, useEffect } from 'react';
 import bootstrap from 'bootstrap/dist/js/bootstrap.bundle.min.js';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './MyYahoo.css';
-
-const fetchData = async (endpoint, setState, timestampKey, prefix, setTimestamps) => {
-  try {
-    const url = `http://localhost:5000${endpoint}`;
-    const response = await fetch(url, { method: 'POST' });
-    const data = await response.json();
-    setState(data);
-
-    setTimestamps((prev) => ({
-      ...prev,
-      [timestampKey]: new Date().toLocaleString(),
-    }));
-  } catch (error) {
-    console.error(`Error fetching ${endpoint}:`, error);
-  }
-};
+import { fetchData } from './api';
+import StockInfo from './components/StockInfo';
+import NewsFeed from './components/NewsFeed';
+import Sports from './components/Sports';
+import WeatherInfo from './components/WeatherInfo';
+import ManageStocksModal from './components/ManageStocksModal';
+import ManageSportsModal from './components/ManageSportsModal';
+import ManageRssModal from './components/ManageRssModal';
+import ManageWeatherModal from './components/ManageWeatherModal';
 
 const App = () => {  
   const [stockInfo, setStockInfo] = useState([]);
@@ -69,15 +62,6 @@ const App = () => {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  useEffect(() => {
-    // Initialize Bootstrap modals
-    const bootstrap = require('bootstrap/dist/js/bootstrap.bundle.min.js');
-    const modalElements = document.querySelectorAll('.modal');
-    modalElements.forEach((modal) => {
-      new bootstrap.Modal(modal);
-    });
-  }, []);
-
   return (
     <div>
       <table style={{ width: '100%' }}>
@@ -92,7 +76,7 @@ const App = () => {
                 width="20px" className="img-link" alt="Manage Stocks" 
               />
             </h2>
-            {buildStockTables(stockInfo, setStockInfo, showTrashIcons, setTimestamps)}
+            <StockInfo stockInfo={stockInfo} showTrashIcons={showTrashIcons} setStockInfo={setStockInfo} setTimestamps={setTimestamps} />
           </td>
           <td valign="top" width="35%" style={{ paddingLeft: '10px' }} rowSpan="2">
             <h2>
@@ -104,7 +88,7 @@ const App = () => {
                 width="20px" className="img-link" alt="Manage Feeds" 
               />
             </h2>
-            {buildNewsTables(newsFeeds, fetchData, setNewsFeeds, showTrashIcons, setTimestamps)}
+            <NewsFeed newsFeeds={newsFeeds} fetchData={fetchData} setNewsFeeds={setNewsFeeds} showTrashIcons={showTrashIcons} setTimestamps={setTimestamps} />
           </td>
           <td valign="top" width="35%" style={{ paddingLeft: '10px' }}>
             <h2>
@@ -122,7 +106,7 @@ const App = () => {
                 {showTrashIcons ? 'Hide Trash Icons' : 'Show Trash Icons'}
               </button>
             </h2>        
-            {buildSportsTables(sportsFeeds)}
+            <Sports sportsFeeds={sportsFeeds} />
             <h2>
               Weather Information - <span className="small-text">{timestamps.weather}</span>
               <img src="refresh.png" alt="Refresh" className="refresh-icon"
@@ -131,927 +115,21 @@ const App = () => {
               <img src="config.png" title="Configure Weather" data-bs-toggle="modal" data-bs-target="#manage-weather" 
                 width="20px" className="img-link" alt="Manage Weather" />
             </h2>
-            {buildWeatherTables(weatherInfo, setWeatherInfo, showTrashIcons, setTimestamps)}
+            <WeatherInfo weatherInfo={weatherInfo} showTrashIcons={showTrashIcons} setWeatherInfo={setWeatherInfo} setTimestamps={setTimestamps} />
           </td>
         </tr>
       </table>
 
       {/* Modals */}
-      <div id="manage-stocks" className="modal">
-        <div className="modal-dialog">
-          <div className="modal-content">
-            <div className="modal-header">
-              <h5 className="modal-title">Add Stocks/Portfolios</h5>
-              <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <div className="modal-body">
-              <form id="add-stock-form">
-                <div className="mb-3">
-                  <label htmlFor="portfolio-name" className="form-label">Portfolio Name</label>
-                  <input id="portfolio-name" className="form-control" type="text" placeholder="Enter Portfolio name" />
-                </div>
-                <div className="mb-3">
-                  <label htmlFor="stock-symbol" className="form-label">Stock Symbols (comma separated)</label>
-                  <input id="stock-symbol" className="form-control" type="text" placeholder="Enter Stock Symbols (e.g., AAPL,GOOGL)" />
-                </div>
-              </form>
-            </div>
-            <div className="modal-footer">
-              <button type="button" className="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-              <button type="button" className="btn btn-primary" onClick={() => addStock(setTimestamps)}>Add Stock</button>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div id="manage-sports" className="modal">
-        <div className="modal-dialog">
-          <div className="modal-content">
-            <div className="modal-header">
-              <h5 className="modal-title">Manage Sports Teams</h5>
-              <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <div className="modal-body">
-              <form id="manage-sports-form">
-                <div className="mb-3">
-                  <label htmlFor="league-selector" className="form-label">Select League</label>
-                  <select id="league-selector" className="form-select" onChange={() => fetchTeamsForLeague()}>
-                    <option value="" disabled selected>-- Select a League --</option>
-                    <option value="NBA">NBA</option>
-                    <option value="NHL">NHL</option>
-                    <option value="MLB">MLB</option>
-                    <option value="NFL">NFL</option>
-                  </select>
-                </div>
-                <div id="team-list-container" className="row">
-                  <p>Select a league to view its teams.</p>
-                </div>
-              </form>
-            </div>
-            <div className="modal-footer">
-              <button type="button" className="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-              <button type="button" className="btn btn-primary" onClick={() => saveSelectedTeams(setTimestamps, setSportsFeeds)}>Save changes</button>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div id="manage-rss" className="modal">
-        <div className="modal-dialog">
-          <div className="modal-content">
-            <div className="modal-header">
-              <h5 className="modal-title">Add RSS Feeds</h5>
-              <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <div className="modal-body">
-              <form id="add-rss-feed-form">
-                <div className="mb-3">
-                  <label htmlFor="rss-category" className="form-label">Category</label>
-                  <input id="rss-category" className="form-control" type="text" placeholder="Enter category name" />
-                </div>
-                <div className="mb-3">
-                  <label htmlFor="rss-name" className="form-label">RSS Feed Name</label>
-                  <input id="rss-name" className="form-control" type="text" placeholder="Enter RSS feed name" />
-                </div>
-                <div className="mb-3">
-                  <label htmlFor="rss-url" className="form-label">RSS Feed URL</label>
-                  <input id="rss-url" className="form-control" type="text" placeholder="Enter RSS feed URL" />
-                </div>
-              </form>
-            </div>
-            <div className="modal-footer">
-              <button type="button" className="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-              <button type="button" className="btn btn-primary" onClick={() => addRSSFeed(setTimestamps, setNewsFeeds)}>Add Feed</button>
-            </div>
-          </div>
-        </div>
-      </div>
-
-      <div id="manage-weather" className="modal">
-        <div className="modal-dialog">
-          <div className="modal-content">
-            <div className="modal-header">
-              <h5 className="modal-title">Add Weather Locations</h5>
-              <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <div className="modal-body">
-              <form id="add-weather-location-form">
-                <table style={{ width: "100%" }}>
-                  <tr style={{ paddingBottom: "12px", height: "48px" }}>
-                    <td style={{ width: "120px", verticalAlign: "middle" }}>
-                      <label htmlFor="weather-days" className="form-label me-2">Days</label>                      
-                    </td>
-                    <td style={{ verticalAlign: "middle" }} colSpan="2">
-                      <input id="weather-days" className="form-control" type="number" min="1" max="7" defaultValue="3" style={{ width: "80px" }} />
-                    </td>
-                  </tr>
-                  <tr style={{ paddingBottom: "12px", height: "48px" }}>
-                    <td style={{ verticalAlign: "middle" }}>
-                      <label htmlFor="weather-location" className="form-label me-2">Location</label>
-                    </td>
-                    <td style={{ verticalAlign: "middle" }} colSpan="2">
-                      <input id="weather-location" className="form-control" type="text" placeholder="Enter city or zip" style={{ flex: 1 }} />                      
-                    </td>
-                  </tr>                  
-                  <tr style={{ paddingBottom: "12px", height: "48px" }}>
-                    <td style={{ verticalAlign: "middle" }}>
-                      <label className="form-check-label" htmlFor="weather-international">International</label>
-                    </td>
-                    <td style={{ verticalAlign: "middle" }}>
-                      <input id="weather-international" className="form-check-input" type="checkbox" /> 
-                    </td>
-                    <td style={{ verticalAlign: "middle" }} align="right">
-                      <button type="button" className="btn btn-secondary ms-2" onClick={() => getLocations()} >Lookup Location</button>
-                    </td>
-                  </tr>
-                </table>               
-                <div id="location-result">
-                  <p>Search for a location first in order to add one!</p>
-                </div>
-              </form>
-            </div>
-            <div className="modal-footer">
-              <button type="button" className="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-              <button
-                type="button"
-                className="btn btn-primary"
-                onClick={() => addWeatherLocation(setWeatherInfo, setTimestamps)}               
-              >
-                Add Location
-              </button>
-            </div>
-          </div>
-        </div>
-      </div>
+      {/* Modals */}
+      <ManageStocksModal setStockInfo={setStockInfo} setTimestamps={setTimestamps} />
+      <ManageSportsModal setSportsFeeds={setSportsFeeds} setTimestamps={setTimestamps} />
+      <ManageRssModal setNewsFeeds={setNewsFeeds} setTimestamps={setTimestamps} />
+      <ManageWeatherModal setWeatherInfo={setWeatherInfo} setTimestamps={setTimestamps} />
       <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
     </div>
   );
 };  
 
-// Updated helper functions
-const buildStockTables = (stockInfo, setStockInfo, showTrashIcons, setTimestamps) => {
-  const removeStock = async (portfolioName, stockSymbol, setTimestamps) => {
-    try {
-      const url = 'http://localhost:5000/removeStock';
-      const response = await fetch(url, {      
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ portfolioName, stockSymbol })
-      });
-
-      const result = await response.json();
-      if (result.success) {
-        alert(result.message);
-        // Refresh the stock data after deletion
-        fetchData('/stockUpdate', setStockInfo, 'stock', 'stock', setTimestamps);
-      } else {
-        alert(result.message);
-      }
-    } catch (error) {
-      console.error('Error removing stock:', error);
-      alert('An error occurred while removing the stock.');
-    }
-  };
-
-  return stockInfo.map((portfolio, index) => (
-    <div key={index}>
-      <div
-        className="data-header"
-        onClick={() => {
-          const tableId = `stock-${index}`;
-          const table = document.getElementById(tableId);
-
-          if (table) {
-            // Toggle the table's visibility
-            const isCollapsed = table.style.display === 'none' ? false : true;
-            table.style.display = isCollapsed ? 'none' : 'table';
-
-            // Update the collapsed state in the database
-            updateCollapsedState('Portfolios', portfolio.name, !isCollapsed);
-          }
-        }}
-      >
-        {portfolio.name}
-      </div>
-      <table
-        id={`stock-${index}`}
-        className="data-table"
-        style={{ display: portfolio.collapsed ? 'none' : 'table' }} // Conditionally hide the table
-      >
-        <thead>
-          <tr>
-            <th>Name</th>
-            <th>Symbol</th>
-            <th>Price</th>
-            <th>Change</th>
-            <th>Change %</th>
-            <th></th>
-          </tr>
-        </thead>
-        <tbody>
-          {portfolio.portfolioData.map((stock, idx) => (
-            <tr key={idx} className={getRowClass(stock.change)}>
-              <td>
-                <a
-                  href={`https://finance.yahoo.com/quote/${stock.symbol}/`}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                >
-                  {stock.name}
-                </a>
-              </td>
-              <td>{stock.symbol}</td>
-              <td>{formatCurrency(stock.price)}</td>
-              <td>{formatCurrency(stock.change || 0)}</td>
-              <td>{stock.changePercent !== undefined ? stock.changePercent.toFixed(2) + '%' : 'N/A'}</td>
-              <td>
-                <span
-                  className="trash-icon"
-                  onClick={() => removeStock(portfolio.name, stock.symbol, setTimestamps)}
-                  style={{
-                    cursor: 'pointer',
-                    color: 'red',
-                    display: showTrashIcons ? 'inline' : 'none',
-                  }}
-                  title="Remove Stock"
-                >
-                  🗑️
-                </span>
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  ));
-};
-
-const buildNewsTables = (newsFeeds, fetchData, setNewsFeeds, showTrashIcons, setTimestamps) => {
-  const removeRSSFeed = async (categoryName, feedName) => {
-    try {
-      const url = 'http://localhost:5000/removeRSSFeed';
-      const response = await fetch(url, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ category: categoryName, feedName }),
-      });
-
-      const result = await response.json();
-      if (result.success) {
-        alert(result.message);
-        // Refresh the news feeds after deletion
-        fetchData('/newsUpdate', setNewsFeeds, 'news', 'news', setTimestamps);
-      } else {
-        alert(result.message);
-      }
-    } catch (error) {
-      console.error('Error removing RSS feed:', error);
-      alert('An error occurred while removing the RSS feed.');
-    }
-  };
-
-  return newsFeeds.map((category, categoryIndex) => (
-    <div key={categoryIndex}>
-      {/* Category Header */}
-      <h3
-        className="data-header"
-        onClick={() => {
-          const categoryContainer = document.getElementById(`category-container-${categoryIndex}`);
-          if (categoryContainer) {
-            // Toggle visibility of all tables in this category
-            const isCollapsed = categoryContainer.style.display === 'none' ? false : true;
-            categoryContainer.style.display = isCollapsed ? 'none' : 'block';
-
-            // Update the collapsed state in the database
-            updateCollapsedState('NewsCategories', category.category, !isCollapsed);
-          }
-        }}
-        style={{ cursor: 'pointer' }}
-      >
-        {category.category}
-      </h3>
-
-      {/* Category Content */}
-      <div id={`category-container-${categoryIndex}`} style={{ display: category.collapsed ? 'none' : 'block' }}>
-        {category.feeds.map((feed, feedIndex) => (
-          <div key={feedIndex}>
-            {/* Feed Table */}
-            <table className="data-table">
-              <thead>
-                <tr className="data-header">
-                  <th
-                    colSpan="1"
-                    onClick={(e) => toggleTableData(e.target, feed.name, category.category)} // Call toggleTableData
-                    style={{ cursor: 'pointer', textAlign: 'left', position: 'relative' }}
-                  >
-                    {feed.name}
-                    <span
-                      className="trash-icon"
-                      onClick={() => removeRSSFeed(category.category, feed.name)}
-                      style={{
-                        cursor: 'pointer',
-                        color: 'red',
-                        display: showTrashIcons ? 'inline' : 'none',
-                        position: 'absolute',
-                        right: '10px',
-                      }}
-                      title="Remove RSS Feed"
-                    >
-                      🗑️
-                    </span>
-                  </th>
-                </tr>
-              </thead>
-              <tbody>
-                {feed.feedData.items.map((item, itemIndex) => (
-                  <tr
-                    key={item.guid || itemIndex}
-                    style={{ display: feed.feedData.collapsed ? 'none' : '' }} // Conditionally hide the row
-                  >
-                    <td colSpan="2">
-                      <a href={item.link} target="_blank" rel="noopener noreferrer">
-                        {item.title}
-                      </a>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-        ))}
-      </div>
-    </div>
-  ));
-};
-
-const buildSportsTables = (sportsFeeds) => {
-  return sportsFeeds
-    .filter(
-      (sport) =>
-        sport.sportData && // Ensure sportData exists
-        sport.sportData.some((data) => data.items && data.items.length > 0) // Filter out sports with no items
-    )
-    .map((sport, index) => (
-      <div key={index}>
-        <div
-          className="data-header"
-          onClick={() => {
-            const tableId = `sports-${index}`;
-            const table = document.getElementById(tableId);
-
-            if (table) {
-              // Toggle the table's visibility
-              const isCollapsed = table.style.display === 'none' ? false : true;
-              table.style.display = isCollapsed ? 'none' : 'table';
-
-              // Update the collapsed state in the database
-              updateCollapsedState('Sports', sport.name, !isCollapsed);
-            }
-          }}
-        >
-          {sport.name}
-        </div>
-        <table
-          id={`sports-${index}`}
-          className="data-table"
-          style={{ display: sport.collapsed ? 'none' : 'table' }} // Conditionally hide the table
-        >
-          <thead>
-            <tr>
-              <th>Date/Event</th>
-              <th>Home</th>
-              <th>Away</th>
-              <th>Status</th>
-              <th>Result</th>
-            </tr>
-          </thead>
-          <tbody>
-            {sport.sportData.map((data, dataIndex) =>
-              data.items.map((item, itemIndex) => (
-                <tr key={`${dataIndex}-${itemIndex}`} className={getRowClass(item.result)}>
-                  <td>
-                    {item.date} /{' '}
-                    <a href={item.link} target="_blank" rel="noopener noreferrer">
-                      {item.title}
-                    </a>
-                  </td>
-                  <td style={{ fontSize: '28px' }} width="112px">
-                    <img src={item.home_logo} alt="Home Logo" width="42" /> {item.home_score}
-                  </td>
-                  <td style={{ fontSize: '28px' }} width="112px">
-                    <img src={item.away_logo} alt="Away Logo" width="42" /> {item.away_score}
-                  </td>
-                  <td>
-                    {item.box_score ? (
-                      <a href={item.box_score} target="_blank" rel="noopener noreferrer">
-                        {item.status}
-                      </a>
-                    ) : (
-                      item.status
-                    )}
-                  </td>
-                  <td>
-                    {item.recap ? (
-                      <a href={item.recap} target="_blank" rel="noopener noreferrer">
-                        {item.result}
-                      </a>
-                    ) : (
-                      item.result
-                    )}
-                  </td>
-                </tr>
-              ))
-            )}
-          </tbody>
-        </table>
-      </div>
-    ));
-};
-
-const buildWeatherTables = (weatherInfo, setWeatherInfo, showTrashIcons, setTimestamps) => {
-  const removeWeatherLocation = async (locationName) => {
-    try {
-      const url = 'http://localhost:5000/removeWeatherLocation';
-      const response = await fetch(url, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ locationName }),
-      });
-
-      const result = await response.json();
-      if (result.success) {
-        alert(result.message);
-        // Refresh the weather data after deletion
-        fetchData('/weatherUpdate', setWeatherInfo, 'weather', 'weather', setTimestamps);
-      } else {
-        alert(result.message);
-      }
-    } catch (error) {
-      console.error('Error removing weather location:', error);
-      alert('An error occurred while removing the weather location.');
-    }
-  };
-
-  return weatherInfo.map((area, index) => (
-    <div key={index}>
-      <div className="data-header">
-        <span
-          onClick={() => {
-            const tableId = `weather-${index}`;
-            const table = document.getElementById(tableId);
-
-            if (table) {
-              // Toggle the table's visibility
-              const isCollapsed = table.style.display === 'none' ? false : true;
-              table.style.display = isCollapsed ? 'none' : 'table';
-
-              // Update the collapsed state in the database
-              updateCollapsedState('Weather', area.name, !isCollapsed);
-            }
-          }}
-          style={{ cursor: 'pointer' }}
-        >
-          {area.name} - {buildDisplay(area.weatherData.current)}
-        </span>
-        <span
-          className="trash-icon ms-2"
-          onClick={() => removeWeatherLocation(area.name)}
-          style={{
-            cursor: 'pointer',
-            color: 'red',
-            display: showTrashIcons ? 'inline' : 'none',
-            position: 'absolute',
-            right: '10px',
-          }}
-          title="Remove Weather Location"
-        >
-          🗑️
-        </span>
-      </div>
-      <table
-        id={`weather-${index}`}
-        className="data-table"
-        style={{ display: area.collapsed ? 'none' : 'table' }} // Conditionally hide the table
-      >
-        <thead>
-          <tr>
-            <th>Date</th>
-            <th>Forecast</th>
-            <th>Temp Range</th>
-            <th>Wind</th>
-            <th>Precipitation</th>
-            <th>Sunshine</th>
-          </tr>
-        </thead>
-        <tbody>
-          {area.weatherData.daily.time.map((time, idx) => (
-            <tr key={idx}>
-              <td>{new Date(time).toLocaleDateString()}</td>
-              <td>{area.weatherData.daily.weatherCode[idx]}</td>
-              <td>
-                {Math.round(area.weatherData.daily.tempMax[idx])}°/
-                {Math.round(area.weatherData.daily.tempMin[idx])}°
-              </td>
-              <td>
-                {degreesToCompass(area.weatherData.daily.windDirection[idx])}{' '}
-                {Math.round(area.weatherData.daily.windSpeed[idx])} MPH
-              </td>
-              <td>{area.weatherData.daily.precipitation[idx]}%</td>
-              <td>{formatSeconds(area.weatherData.daily.sun[idx])}</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </div>
-  ));
-};
-
-// Helper functions
-const getRowClass = (value) => {
-
-  let rowClass = 'row-neutral';
-  if (value === "Win")
-    rowClass = 'row-positive';
-  else if (value === "Loss")
-    rowClass = 'row-negative';
-
-  return rowClass;
-};
-
-const formatCurrency = (value) => {
-  return new Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' }).format(value);
-};
-
-const degreesToCompass = (degrees) => {
-  const directions = ['N', 'NNE', 'NE', 'ENE', 'E', 'ESE', 'SE', 'SSE', 'S', 'SSW', 'SW', 'WSW', 'W', 'WNW', 'NW', 'NNW'];
-  const index = Math.round(degrees / 22.5) % 16;
-  return directions[index];
-};
-
-const formatSeconds = (totalSeconds) => {
-  totalSeconds = Math.round(totalSeconds);
-
-  const hours = Math.floor(totalSeconds / 3600);
-  const minutes = Math.floor((totalSeconds % 3600) / 60);
-  const remainingSeconds = totalSeconds % 60;
-
-  const formattedHours = String(hours).padStart(2, '0');
-  const formattedMinutes = String(minutes).padStart(2, '0');
-  const formattedSeconds = String(remainingSeconds).padStart(2, '0');
-  
-  if (hours > 0) {
-      return `${formattedHours}:${formattedMinutes}:${formattedSeconds}`;
-  } else {
-      return `${formattedMinutes}:${formattedSeconds}`;
-  }
-};
-
-const buildDisplay = (area) => {
-    let display = `${area.weatherCode} ${Math.round(area.temp)}° Feels like: ${Math.round(area.feelsLike)}°`;
-    display += ` Humidity: ${area.humidty}%`
-    if (area.windSpeed > 9)
-        display += `, ${degreesToCompass(area.windDirection)} ${Math.round(area.windSpeed)}MPH`;
-    return display;
-}
-
-const addStock = async (setTimestamps) => {
-  const portfolioName = document.getElementById('portfolio-name').value.trim();
-  const stockSymbol = document.getElementById('stock-symbol').value.trim();
-  const setStockInfo = (data) => console.log('setStockInfo is not defined in this scope. Replace this with actual implementation.');
-
-  if (!portfolioName || !stockSymbol) {
-    alert('Please enter both a portfolio name and stock symbols.');
-    return;
-  }
-
-  try {
-    const url = 'http://localhost:5000/addStock';
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json'},
-      body: JSON.stringify({ portfolioName, stockSymbol })
-    });
-
-    const result = await response.json();
-    if (result.success) {
-      alert(result.message);
-      // Refresh the stock data after adding
-      fetchData('/stockUpdate', setStockInfo, 'stock', 'stock', setTimestamps);
-
-      // Close the modal
-      const modalElement = document.getElementById('manage-stocks');
-      const modalInstance = bootstrap.Modal.getInstance(modalElement);
-      modalInstance.hide();
-    } else {
-      alert(result.message);
-    }
-  } catch (error) {
-    console.error('Error adding stock:', error);
-    alert('An error occurred while adding the stock.');
-  }
-};
-
-const addRSSFeed = async (setTimestamps, setNewsFeeds) => {
-  const category = document.getElementById('rss-category').value.trim();
-  const name = document.getElementById('rss-name').value.trim();
-  const rssUrl = document.getElementById('rss-url').value.trim();
-
-  if (!category || !name || !rssUrl) {
-    alert('Please fill out all fields: Category, RSS Feed Name, and RSS Feed URL.');
-    return;
-  }
-
-  try {
-    const url = 'http://localhost:5000/addRSSFeed';
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ category, name, rssUrl })
-    });
-
-    const result = await response.json();
-    if (result.success) {
-      alert(result.message);
-      // Refresh the news feeds after adding
-      fetchData('/newsUpdate', setNewsFeeds, 'news', 'news', setTimestamps);
-    
-      // Close the modal
-      const modalElement = document.getElementById('manage-rss');
-      const modalInstance = bootstrap.Modal.getInstance(modalElement);
-      modalInstance.hide();
-    } else {
-      alert(result.message);
-    }
-  } catch (error) {
-    console.error('Error adding RSS feed:', error);
-    alert('An error occurred while adding the RSS feed.');
-  }
-};
-
-const getLocations = async () => {
-  const location = document.getElementById('weather-location').value.trim();
-  if (!location) {
-    alert('Please fill in the location field.');
-    return;
-  }
-
-  const international = document.getElementById('weather-international').checked;
-  
-  // Clear previous results
-  const locationResultElement = document.getElementById('location-result');
-  locationResultElement.innerHTML = '';
-
-  try {
-    // Call the external REST endpoint
-    const url = 'http://localhost:5000/searchLocation';
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ locationName: location, international }),
-    });
-
-    const result = await response.json();
-    if (result.success) {
-      // Populate the location-result field with a list of radio buttons
-      const resultList = document.createElement('div');
-      resultList.className = 'list-group';
-
-      result.locations.forEach((location, index) => {
-        const radioItem = document.createElement('div');
-        radioItem.className = 'form-check';
-
-        const input = document.createElement('input');
-        input.className = 'form-check-input';
-        input.type = 'radio';
-        input.name = 'locationRadio';
-        input.id = `location-${index}`;
-        input.value = `${location.name}, ${location.admin1 || ''} | ${location.latitude} | ${location.longitude}`;
-
-        const label = document.createElement('label');
-        label.className = 'form-check-label';
-        label.htmlFor = `location-${index}`;
-        label.textContent = `${location.name}, ${location.admin1 || ''}, ${location.country} (Lat: ${location.latitude}, Lon: ${location.longitude})`;
-
-        radioItem.appendChild(input);
-        radioItem.appendChild(label);
-        resultList.appendChild(radioItem);
-      });
-
-      locationResultElement.appendChild(resultList);
-    } else {
-      locationResultElement.textContent = result.message;
-    }
-  } catch (error) {
-    console.error('Error searching for locations:', error);
-    locationResultElement.textContent = 'An error occurred while searching for locations.';
-  }
-};
-
-const addWeatherLocation = async (setWeatherInfo, setTimestamps) => {
-  // Get the selected location from the radio buttons
-  const selectedLocation = document.querySelector('input[name="locationRadio"]:checked');
-  if (!selectedLocation) {
-    alert('Please select a location to add.');
-    return;
-  }
-
-  // Extract location details from the selected radio button's value
-  const [locationName, latitude, longitude] = selectedLocation.value.split('|').map((value) => value.trim());
-  const days = document.getElementById('weather-days').value.trim();
-
-  if (!locationName || !latitude || !longitude) {
-    alert('Invalid location data. Please try again.');
-    return;
-  }
-
-  try {
-    // Send the location data to the server
-    const url = 'http://localhost:5000/addWeatherLocation';
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ locationName, latitude, longitude, days }),
-    });
-    const result = await response.json();
-
-    if (result.success) {
-      alert(result.message);
-      // Refresh the weather data after adding the location
-      fetchData('/weatherUpdate', setWeatherInfo, 'weather', 'weather', setTimestamps);
-
-      // Close the modal
-      const modalElement = document.getElementById('manage-weather');
-      const modalInstance = bootstrap.Modal.getInstance(modalElement);
-      modalInstance.hide();
-    } else {
-      alert(result.message);
-    }
-  } catch (error) {
-    console.error('Error adding weather location:', error);
-    alert('An error occurred while adding the weather location.');
-  }
-};
-
-const fetchTeamsForLeague = async () => {
-  const leagueSelector = document.getElementById('league-selector');
-  const selectedLeague = leagueSelector.value;
-
-  if (!selectedLeague) {
-    alert('Please select a league.');
-    return;
-  }
-
-  try {
-    // Call the server-side getTeamList method
-    const url = "http://localhost:5000/getTeamList";
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ league: selectedLeague }),
-    });
-    const responseData = await response.json();
-
-    const teamListContainer = document.getElementById('team-list-container');
-    teamListContainer.innerHTML = ''; // Clear previous results
-
-    if (responseData.success) {
-      // Create a row container for two-column layout
-      var row = document.createElement('div');
-      row.className = 'row';
-
-      responseData.teams.forEach((team, index) => {
-        // Create a column for each team
-        const col = document.createElement('div');
-        col.className = 'col-md-6'; // Two columns per row
-
-        const teamItem = document.createElement('div');
-        teamItem.className = 'form-check d-flex align-items-center';
-
-        const input = document.createElement('input');
-        input.className = 'form-check-input me-2';
-        input.type = 'checkbox';
-        input.id = `team-${team.id}`;
-        input.value = `${team.id}|${team.logo}`;;
-        input.checked = team.selected;
-
-        const logo = document.createElement('img');
-        logo.src = team.logo; // Use the team logo URL
-        logo.alt = `${team.name} Logo`;
-        logo.style.width = '30px';
-        logo.style.height = '30px';
-        logo.className = 'me-2';
-
-        const label = document.createElement('label');
-        label.className = 'form-check-label';
-        label.htmlFor = `team-${team.id}`;
-        label.textContent = team.name;
-
-        // Append elements to the team item
-        teamItem.appendChild(input);
-        teamItem.appendChild(logo);
-        teamItem.appendChild(label);
-
-        // Append the team item to the column
-        col.appendChild(teamItem);
-
-        // Append the column to the row
-        row.appendChild(col);
-
-        // Add a new row after every two columns
-        if ((index + 1) % 2 === 0) {
-          teamListContainer.appendChild(row);
-          row = document.createElement('div');
-          row.className = 'row';
-        }
-      });
-
-      // Append the last row if it has remaining columns
-      if (row.children.length > 0) {
-        teamListContainer.appendChild(row);
-      }
-    } else {
-      teamListContainer.textContent = responseData.message;
-    }
-  } catch (error) {
-    console.error('Error fetching teams:', error);
-    const teamListContainer = document.getElementById('team-list-container');
-    teamListContainer.textContent = `An error occurred while fetching teams. ${error}`;
-  }
-};
-
-const saveSelectedTeams = async (setTimestamps, setSportsFeeds) => {
-  const leagueSelector = document.getElementById('league-selector');
-  const selectedLeague = leagueSelector.value;
-
-  if (!selectedLeague) {
-    alert('Please select a league.');
-    return;
-  }
-
-  const selectedTeams = [];
-  document.querySelectorAll('#team-list-container .form-check-input:checked').forEach(checkbox => {
-    const [teamId, teamLogo] = checkbox.value.split('|');
-    const teamName = document.querySelector(`label[for="team-${teamId}"]`).textContent.trim();
-    selectedTeams.push({ id: teamId, logo: teamLogo, name: teamName });
-  });
-
-  if (selectedTeams.length === 0) {
-    alert('No teams selected to save.');
-    return;
-  }
-
-  try {
-    const url = 'http://localhost:5000/saveSelectedTeams';
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ league: selectedLeague, selectedTeams }),
-    });
-    const result = await response.json();
-    if (result.success) {
-      alert(result.message);
-      fetchData('/sportsUpdate', setSportsFeeds, 'sports', 'sports', setTimestamps);
-      
-      // Close the modal
-      const modalElement = document.getElementById('manage-sports');
-      const modalInstance = bootstrap.Modal.getInstance(modalElement);
-      modalInstance.hide();
-    } else {
-      alert(result.message);
-    }
-  } catch (error) {
-    console.error('Error saving selected teams:', error);
-    alert('An error occurred while saving the selected teams.');
-  }
-};
-
-        // Hide the table when clicking the header
-function toggleTableData(header, name, category) {
-  const table = header.closest('table');
-  const alltr = table.querySelectorAll("tr");
-  const isCollapsed = alltr[1].style.display === "none" ? true : false; 
-  for (var i = 1; i < alltr.length; i++) {
-      alltr[i].style.display = (alltr[i].style.display === "none") ? "" : "none";
-  }
-  
-  updateCollapsedState("NewsFeeds", category + ":" + name, isCollapsed); // Update the collapsed state in the database            
-}  
-
-const updateCollapsedState = async (section, name, collapsed) => {
-  try {
-    const url = 'http://localhost:5000/updateCollapsedState';
-    const response = await fetch(url, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({ section, name, collapsed: !collapsed }),
-    });
-    const result = await response.json();
-    if (!result.success) {
-      alert(result.message);
-    }
-  } catch (error) {
-    alert('An error occurred while updating the collapsed state.');
-  }
-};
  
 export default App;
